Joseph-AI-Code2-main-main/requirements.txt
# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0

# Utilities
Pillow>=10.0.0
python-dateutil>=2.8.2

# Optional: for production
gunicorn>=21.0.0
whitenoise>=6.6.0
